Renderer: satisfy the cppcheck and MSVC lanes on the RHI boundary
Name the created program before returning it so cppcheck no longer ties
the handle's lifetime to the variant source buffers, and cast the slot
index explicitly for the 32-bit free-list entry (MSVC C4267).

+++ renderer/src/device_slot_table.h
@@ -70,7 +70,7 @@ public:
     entry.live = false;
     entry.payload = Payload{};
     entry.generation = next_generation(entry.generation);
-    m_freeList[m_freeCount] = slot;
+    m_freeList[m_freeCount] = static_cast<std::uint32_t>(slot);
     ++m_freeCount;
     return true;
   }
